# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: XDIGITEX AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List agents
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create/spawn an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Stop and delete an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /agents/{id}/timeline:
    get:
      operationId: getAgentTimeline
      tags: [agents]
      summary: Get agent timeline events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
...
  /apikeys:
    get:
      operationId: listApiKeys
      tags: [apikeys]
      summary: List API keys
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [apikeys]
      summary: Create an API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
  /apikeys/{id}:
    delete:
      operationId: deleteApiKey
      tags: [apikeys]
      summary: Delete an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ADMIN
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/status:
    patch:
      operationId: adminUpdateUserStatus
      tags: [admin]
      summary: Admin - update user status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
  /admin/providers:
    get:
      operationId: adminListProviders
      tags: [admin]
      summary: Admin - list AI providers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiProvider"
    post:
      operationId: adminCreateProvider
      tags: [admin]
      summary: Admin - add AI provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiProviderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
  /admin/providers/{id}:
    patch:
      operationId: adminUpdateProvider
      tags: [admin]
      summary: Admin - update AI provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiProviderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
    delete:
      operationId: adminDeleteProvider
      tags: [admin]
      summary: Admin - delete AI provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/routing:
    get:
      operationId: adminGetRouting
      tags: [admin]
      summary: Admin - get model routing rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRoute"
    put:
      operationId: adminUpdateRouting
      tags: [admin]
      summary: Admin - update model routing rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRoutingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRoute"
...
  /admin/audit-logs:
    get:
      operationId: adminGetAuditLogs
      tags: [admin]
      summary: Admin - get audit logs
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, moderator, support, user]
        status:
          type: string
          enum: [active, suspended, banned]
        avatarUrl:
          type: ["string", "null"]
        plan:
          type: string
        createdAt:
          type: string
          format: date-time

    # DASHBOARD
...
    Agent:
      type: object
      required: [id, name, type, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [planner, architect, frontend, backend, devops, qa, security, reviewer, research, telegram_bot_builder]
        status:
          type: string
          enum: [idle, running, completed, failed, paused]
        projectId:
          type: ["integer", "null"]
        model:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        task:
          type: ["string", "null"]
        progress:
          type: integer
        runtime:
          type: integer
        createdAt:
          type: string
          format: date-time
    AgentInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        projectId:
          type: integer
        task:
          type: string
        model:
          type: string
    TimelineEvent:
      type: object
      required: [id, agentId, action, description, timestamp]
      properties:
        id:
          type: integer
        agentId:
          type: integer
        action:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [success, failed, in_progress]
        timestamp:
          type: string
          format: date-time
    AgentRoute:
      type: object
      required: [agentType, provider, model]
      properties:
        agentType:
          type: string
        provider:
          type: string
        model:
          type: string
    AgentRoutingUpdate:
      type: object
      required: [routes]
      properties:
        routes:
          type: array
          items:
            $ref: "#/components/schemas/AgentRoute"

    # BOTS
...
    ApiKey:
      type: object
      required: [id, name, prefix, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        prefix:
          type: string
        lastUsed:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ApiKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ApiKeyCreated:
      type: object
      required: [id, name, prefix, key, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        prefix:
          type: string
        key:
          type: string
        createdAt:
          type: string
          format: date-time

    # ADMIN
    AdminUser:
      type: object
      required: [id, name, email, role, status, plan, spend, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        plan:
          type: string
        spend:
          type: number
        createdAt:
          type: string
          format: date-time
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, suspended, banned]
...
    AiProvider:
      type: object
      required: [id, name, type, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [active, inactive]
        keyCount:
          type: integer
        modelCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    AiProviderInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
    AiProviderUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
...
    AuditLog:
      type: object
      required: [id, user, action, resource, timestamp]
      properties:
        id:
          type: integer
        user:
          type: string
        action:
          type: string
        resource:
          type: string
        details:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
...